+++ src/app/shared/shared.module.ts
@@ -13,6 +13,7 @@ import { FabSpeedDialModule } from './modules/fab-speed-dial/fab-speed-dial.modu
 import { CathPci50ListControlComponent } from '../feature/registry/cath-pci50-list-control/cath-pci50-list-control.component';
 import { ACSx290ListControlComponent } from '../feature/registry/acsx290-list-control/acsx290-list-control.component';
 import { IsoDatePipe } from './pipes/iso-date.pipe';
+import { ChartsModule } from 'ng2-charts';
 
 @NgModule({
   declarations: [
@@ -30,7 +31,8 @@ import { IsoDatePipe } from './pipes/iso-date.pipe';
     MaterialModule,
     FlexLayoutModule,
     RegistryFormModule,
-    FabSpeedDialModule
+    FabSpeedDialModule,
+    ChartsModule
   ],
   exports: [
     CommonModule,
@@ -44,7 +46,8 @@ import { IsoDatePipe } from './pipes/iso-date.pipe';
     IsoDatePipe,
     FabSpeedDialModule,
     CathPci50ListControlComponent,
-    ACSx290ListControlComponent
+    ACSx290ListControlComponent,
+    ChartsModule
   ],
   providers: [],
   entryComponents: [ModalDialogComponent]